:recycle: Refactor: 코드 정리 및 css 중복 제거
