+++ src/components/LCK/components/Match/Match.tsx
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import { useState } from 'react';
 import styles from './Match.module.scss';
 import Schedule from './Schedule/Schedule';
+
 const Match = () => {
   const [isHover, setIsHover] = useState<boolean>(false);
   const handleClick = () => {
