+++ src/components/LCK/components/Match/Schedule/Item/ItenResponsive.module.scss
@@ -4,7 +4,6 @@
       width: 100%;
       .match_container {
         display: flex;
-        flex-direction: row;
         justify-content: center;
         width: 100%;
         .match_info {
