+++ src/components/LCK/components/Match/Schedule/Month/Month.tsx
@@ -3,7 +3,6 @@ import { useDateStore } from '@/components/LCK/Zustand/myMonth';
 import React, { useCallback, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styles from './Month.module.scss';
-import { useTeams } from '@/components/LCK/Zustand/useTeams';
 
 const Month = () => {
   const [month, setMonth] = useState<any[]>([]);
