+++ src/components/LCK/components/Match/Schedule/Sidebar/SideBar.module.scss
@@ -1,11 +1,10 @@
 .sidebar_container {
   margin: 0;
-  position: absolute;
+  position: sticky;
   right: 0;
+  width: 28%;
   min-height: 100vh;
-  background: inherit;
   transition: all 0.5s ease;
-  background-color: #2f2f2f;
   box-shadow: 1px 1px 1px rgba(200, 200, 200, 0.6);
 }
